requirements.txt
django
djangorestframework
djangorestframework-simplejwt
gunicorn
drf-yasg
psycopg2-binary
django-cors-headers
Pillow
django-cloudinary-storage
cloudinary
pillow